Restructure the packages
Have packages for different layers
