+++ src/main/java/team/yama/yama/repository/UserRepository.java
@@ -1,6 +1,7 @@
-package team.yama.yama.user;
+package team.yama.yama.repository;
 
 import org.springframework.data.repository.CrudRepository;
+import team.yama.yama.domain.User;
 
 public interface UserRepository extends CrudRepository<User, Long> {
 }
